Move failing tests (wrt #192) under `tofix` for now
